double四則運算時有誤差，改decimal

+++ count2/Program.cs
@@ -7,7 +7,8 @@ namespace count2
     {
         static void Main(string[] args)
         {
-            double num1, num2;
+            //string時，雖然成功回傳null及數值，中間轉換double四則運算時有誤差，改decimal
+            decimal num1, num2;
             var val1 = "";
             var val2 = "";
             var inputVal1 = 1;
@@ -21,8 +22,9 @@ namespace count2
             Console.WriteLine("===============\n");
             val2 = InputVal.GetInt(inputVal2);
 
-            num1 = double.Parse(val1);
-            num2 = double.Parse(val2);
+            //string時，雖然成功回傳null及數值，中間轉換double四則運算時有誤差，改decimal
+            num1 = decimal.Parse(val1);
+            num2 = decimal.Parse(val2);
             Console.WriteLine("===============\n");
             Console.WriteLine("請輸入運算符號(+,-,*,/): ");
             var oper = Console.ReadLine();
